Added more buttons, including log out button. Fixed navbar alignment.
