+++ src/App.css
@@ -31,6 +31,7 @@ background-color: black;
 
 li{
   display:inline-block;
+  margin-left: auto;
   margin: auto;
   font-size: 20px;
   padding: 15px;
